Add ms support
Milliseconds are the core format of date handling in javascript, so
having the offset in ms is very useful.

Previously the module exposed an object with a `minutes()` and an
`hours()` function. It now exposes a function which returns offset in
ms. On that function is two methods, `minutes()` and `hours()` which
work as before. This way it's completely backwards compatible + they are
pretty convenient